+++ docs/src/changelog.md
@@ -17,6 +17,7 @@ If your currently installed *Quality-time* version is not the latest version, pl
 ### Added
 
 - When measuring source-up-to-dateness with Dependency-Track as source, also show the up-to-dateness of individual projects in the measurement details. Closes [#10545](https://github.com/ICTU/quality-time/issues/10545).
+- When measuring security warnings or dependencies with Dependency-Track as source, allow for including only project versions that are the latest version. Closes [#11121](https://github.com/ICTU/quality-time/issues/11121).
 
 ### Fixed